+++ include/frontend/Tokenizer.h
@@ -24,7 +24,9 @@ namespace panther{
 			EVO_NODISCARD auto tokenize_comment() noexcept -> bool;
 			EVO_NODISCARD auto tokenize_ident() noexcept -> bool;
 			EVO_NODISCARD auto tokenize_punctuation() noexcept -> bool;
+			EVO_NODISCARD auto tokenize_number_literal() noexcept -> bool;
 			EVO_NODISCARD auto tokenize_string_literal() noexcept -> bool;
+
 
 
 		
@@ -36,12 +38,18 @@ namespace panther{
 
 				union Value{
 					size_t index;
+
 					bool boolean;
+
+					uint64_t integer;
+					long double floating_point;
+
 					// might be undef
 				} value;
 
 				explicit TokenData(Token::Kind token_kind, size_t index_val) : kind(token_kind), value(index_val) {};
 				explicit TokenData(Token::Kind token_kind, bool bool_val) : kind(token_kind), value(bool_val) {};
+				explicit TokenData(Token::Kind token_kind, long double double_val) : kind(token_kind) { this->value.floating_point = double_val; };
 				explicit TokenData(Token::Kind token_kind) : kind(token_kind) {};
 			};
 
@@ -63,9 +71,12 @@ namespace panther{
 
 
 			EVO_NODISCARD auto getStringValue(TokenID id) const noexcept -> const std::string&;
+			EVO_NODISCARD auto getBoolValue(TokenID id) const noexcept -> bool;
+			EVO_NODISCARD auto getIntegerValue(TokenID id) const noexcept -> uint64_t;
 
 
 			auto print() const noexcept -> void;
+
 
 		private:
 			Tokenizer& tokenizer;